+++ ui/option.c
@@ -36,7 +36,7 @@ t_opts opts = { .target = NULL, .range = {.min = MINTTL - 1, .max = MAXTTL},
 };
 
 typedef struct ent_rad_map {
-  guint ndx;
+  uint ndx;
   int   val;
   char *str;
 } t_ent_rad_map;
@@ -59,6 +59,7 @@ t_ent_bool ent_bool[] = {
   [ENT_BOOL_CC]   = { .en.type = ENT_BOOL_CC,   .valfn = pingelem_enabler, .valtype = PE_CC   },
   [ENT_BOOL_DESC] = { .en.type = ENT_BOOL_DESC, .valfn = pingelem_enabler, .valtype = PE_DESC },
   [ENT_BOOL_RT]   = { .en.type = ENT_BOOL_RT,   .valfn = pingelem_enabler, .valtype = PE_RT   },
+  [ENT_BOOL_WMF]  = { .en.type = ENT_BOOL_WMF,  .pval = &opts.whois_multi },
   [ENT_BOOL_LOSS] = { .en.type = ENT_BOOL_LOSS, .valfn = pingelem_enabler, .valtype = PE_LOSS },
   [ENT_BOOL_SENT] = { .en.type = ENT_BOOL_SENT, .valfn = pingelem_enabler, .valtype = PE_SENT },
   [ENT_BOOL_RECV] = { .en.type = ENT_BOOL_RECV, .valfn = pingelem_enabler, .valtype = PE_RECV },
@@ -109,6 +110,7 @@ t_ent_str ent_str[] = {
 
 static t_ent_exp ent_exp[] = {
   [ENT_EXP_INFO] = { .c.en.type = ENT_EXP_INFO },
+  [ENT_EXP_WMF]  = { .c.en.type = ENT_EXP_WMF  }, // adjacent to ENT_EXP_INFO
   [ENT_EXP_STAT] = { .c.en.type = ENT_EXP_STAT },
   [ENT_EXP_LGFL] = { .c.en.type = ENT_EXP_LGFL },
   [ENT_EXP_GREX] = { .c.en.type = ENT_EXP_GREX },
@@ -215,31 +217,33 @@ static t_ent_rad ent_rad[] = {
 };
 
 static gboolean opt_silent;
-#define OPT_NOTIFY(...) do { if (!opt_silent) notifier_inform(__VA_ARGS__); } \
-  while (0)
+#define OPT_NOTIFY(...) do { if (!opt_silent) notifier_inform(__VA_ARGS__); } while (0)
 
 //
 
-static gboolean check_bool_val(GtkCheckButton *check, t_ent_bool *en, void (*update)(void)) {
-  gboolean state = gtk_check_button_get_active(check), okay = false;
-  if (en) {
-    gboolean *pbool = EN_BOOLPTR(en);
-    if (pbool && (*pbool != state)) {
-      *pbool = state; okay = true;
-      if (update) update();
-    }
-    const char *name = en->en.name;
-    if (name && (*name == '_'))
-      name++;
-    const char *enabled = state ? ON_HDR : OFF_HDR;
-    if (en->prefix) OPT_NOTIFY("%s: %s: %s", en->prefix, name, enabled);
-    else            OPT_NOTIFY("%s: %s",                 name, enabled);
+static gboolean check_bool_val(GtkCheckButton *check, t_ent_bool *en, void (*update)(void)) { // NONNULL(1, 2, 3)
+  // check is GtkCheckButton
+  gboolean *pbool = EN_BOOLPTR(en);
+  gboolean state = gtk_check_button_get_active(check);
+  gboolean changed = pbool && (*pbool != state);
+  if (changed) {
+    *pbool = state;
+    update();
   }
-  return okay;
+  const char *name = en->en.name;
+  if (name && (*name == '_'))
+    name++;
+  const char *enabled = state ? ON_HDR : OFF_HDR;
+  if (en->prefix)
+    OPT_NOTIFY("%s: %s: %s", en->prefix, name, enabled);
+  else
+    OPT_NOTIFY("%s: %s",                 name, enabled);
+  return changed;
 }
 
 static void set_ed_texthint(t_ent_str *en) {
-  if (!en) return;
+  if (!en)
+    return;
   g_return_if_fail(GTK_IS_EDITABLE(en->input));
   int  *pint = en->pint;
   char *pstr = en->pstr;
@@ -260,92 +264,122 @@ static void toggled_dns(void) {
   pinger_update_tabs(NULL);
 }
 
-static void toggle_colors(void) {
+static void toggled_whois(void) {
+  pingtab_vis_cols();
+  notifier_legend_vis_rows(-1);
+  stat_whois_enabler();
+  if (opts.whois)
+    stat_run_whois_resolv();
+}
+
+static void toggled_mwhois(void) {
+  notifier_legend_vis_rows(-1);
+  if (opts.whois) {
+    stat_whois_review(parser_review_whois);
+    pingtab_update();
+  }
+}
+
+static void toggled_colors(void) {
   tab_dependent(NULL); style_set();
   tab_reload_theme();  drawtab_refresh();
   notifier_legend_refresh();
 }
 
-static void toggle_legend(void) { if (opts.graph) notifier_set_visible(NT_LEGEND_NDX, opts.legend); graphtab_refresh(); }
+static void toggled_legend(void) {
+  if (opts.graph)
+    notifier_set_visible(NT_LEGEND_NDX, opts.legend);
+  graphtab_refresh();
+}
+
+static void toggled_graph_params(void) {
+  notifier_legend_vis_rows(-1);
+  graphtab_refresh();
+}
+
+static void toggled_graph_area(void) {
+  if (opts.legend)
+    notifier_legend_update();
+  graphtab_refresh();
+}
+
 #ifdef WITH_PLOT
-static void toggle_pl_elems(void) { plottab_refresh(PL_PARAM_AT); }
-static void toggle_pl_rotor(void) { notifier_set_visible(NT_ROTOR_NDX, is_plelem_enabled(D3_ROTR)); toggle_pl_elems(); }
+static void toggled_pl_elems(void) {
+  plottab_refresh(PL_PARAM_AT);
+}
+
+static void toggled_pl_rotor(void) {
+  notifier_set_visible(NT_ROTOR_NDX, is_plelem_enabled(D3_ROTR));
+  toggled_pl_elems();
+}
 #endif
 
+typedef void (*toggled_bool_fn)(void); // GVoidFunc
+
 static void toggle_cb(GtkCheckButton *check, t_ent_bool *en) {
-  if (!GTK_IS_CHECK_BUTTON(check) || !en) return;
-  switch (en->en.type) {
-    case ENT_BOOL_DNS:
-      check_bool_val(check, en, toggled_dns);
-      break;
-    case ENT_BOOL_HOST:
-    case ENT_BOOL_LOSS:
-    case ENT_BOOL_SENT:
-    case ENT_BOOL_RECV:
-    case ENT_BOOL_LAST:
-    case ENT_BOOL_BEST:
-    case ENT_BOOL_WRST:
-    case ENT_BOOL_AVRG:
-    case ENT_BOOL_JTTR:
-      check_bool_val(check, en, pingtab_vis_cols);
-      break;
-    case ENT_BOOL_AS:
-    case ENT_BOOL_CC:
-    case ENT_BOOL_DESC:
-    case ENT_BOOL_RT:
-      if (check_bool_val(check, en, pingtab_vis_cols)) notifier_legend_vis_rows(-1);
-      stat_whois_enabler();
-      if (opts.whois) stat_run_whois_resolv();
-      break;
-    case ENT_BOOL_MN_DARK:
-    case ENT_BOOL_GR_DARK:
+  static toggled_bool_fn toggled_bool[] = {
+    [ENT_BOOL_DNS]     = toggled_dns,
+    [ENT_BOOL_HOST]    = pingtab_vis_cols,
+    [ENT_BOOL_LOSS]    = pingtab_vis_cols,
+    [ENT_BOOL_SENT]    = pingtab_vis_cols,
+    [ENT_BOOL_RECV]    = pingtab_vis_cols,
+    [ENT_BOOL_LAST]    = pingtab_vis_cols,
+    [ENT_BOOL_BEST]    = pingtab_vis_cols,
+    [ENT_BOOL_WRST]    = pingtab_vis_cols,
+    [ENT_BOOL_AVRG]    = pingtab_vis_cols,
+    [ENT_BOOL_JTTR]    = pingtab_vis_cols,
+    [ENT_BOOL_AS]      = toggled_whois,
+    [ENT_BOOL_CC]      = toggled_whois,
+    [ENT_BOOL_DESC]    = toggled_whois,
+    [ENT_BOOL_RT]      = toggled_whois,
+    [ENT_BOOL_WMF]     = toggled_mwhois,
+    [ENT_BOOL_MN_DARK] = toggled_colors,
+    [ENT_BOOL_GR_DARK] = toggled_colors,
 #ifdef WITH_PLOT
-    case ENT_BOOL_PL_DARK:
+    [ENT_BOOL_PL_DARK] = toggled_colors,
 #endif
-      check_bool_val(check, en, toggle_colors);
-      break;
-    case ENT_BOOL_LGND:
-      check_bool_val(check, en, toggle_legend);
-      break;
-    case ENT_BOOL_AVJT:
-    case ENT_BOOL_CCAS:
-    case ENT_BOOL_LGHN:
-      if (check_bool_val(check, en, NULL)) { notifier_legend_vis_rows(-1); graphtab_refresh(); }
-      break;
-    case ENT_BOOL_MEAN:
-    case ENT_BOOL_JRNG:
-    case ENT_BOOL_AREA:
-      if (check_bool_val(check, en, NULL)) { if (opts.legend) notifier_legend_update(); graphtab_refresh(); }
-      break;
+    [ENT_BOOL_LGND]    = toggled_legend,
+    [ENT_BOOL_AVJT]    = toggled_graph_params,
+    [ENT_BOOL_CCAS]    = toggled_graph_params,
+    [ENT_BOOL_LGHN]    = toggled_graph_params,
+    [ENT_BOOL_MEAN]    = toggled_graph_area,
+    [ENT_BOOL_JRNG]    = toggled_graph_area,
+    [ENT_BOOL_AREA]    = toggled_graph_area,
 #ifdef WITH_PLOT
-    case ENT_BOOL_PLBK:
-    case ENT_BOOL_PLAX:
-    case ENT_BOOL_PLGR:
-      check_bool_val(check, en, toggle_pl_elems);
-      break;
-    case ENT_BOOL_PLRR:
-      check_bool_val(check, en, toggle_pl_rotor);
-      break;
-    case ENT_BOOL_GLOB:
-      check_bool_val(check, en, option_up_menu_ext);
-      break;
+    [ENT_BOOL_PLBK]    = toggled_pl_elems,
+    [ENT_BOOL_PLAX]    = toggled_pl_elems,
+    [ENT_BOOL_PLGR]    = toggled_pl_elems,
+    [ENT_BOOL_PLRR]    = toggled_pl_rotor,
+    [ENT_BOOL_GLOB]    = option_up_menu_ext,
 #endif
-    default: break;
+  };
+  if (GTK_IS_CHECK_BUTTON(check) && en) {
+    int type = en->en.type;
+    toggled_bool_fn fn = (type >= 0) && (type < (int)G_N_ELEMENTS(toggled_bool)) ?
+      toggled_bool[type] : NULL;
+    if (fn)
+      check_bool_val(check, en, fn);
   }
 }
 
 static void radio_cb(GtkCheckButton *check, t_ent_rad_map *map) {
-  if (!GTK_IS_CHECK_BUTTON(check) || !map) return;
-  guint ndx = map->ndx; if (ndx >= G_N_ELEMENTS(ent_rad)) return;
+  if (!GTK_IS_CHECK_BUTTON(check) || !map)
+    return;
+  uint ndx = map->ndx;
+  if (ndx >= G_N_ELEMENTS(ent_rad))
+    return;
   t_ent_rad *en = &ent_rad[ndx];
   int *pval = en->pval, type = en->c.en.type;
   if (pval && ((type == ENT_RAD_IPV) || (type == ENT_RAD_GRAPH))) {
     gboolean selected = gtk_check_button_get_active(check);
     if (selected && (map->val != *pval)) {
       *en->pval = map->val;
-      if (en->cb) en->cb();
-      if (map->str) OPT_NOTIFY("%s: %s", en->c.en.name, map->str);
-      else          OPT_NOTIFY("%s: %d", en->c.en.name, map->val);
+      if (en->cb)
+        en->cb();
+      if (map->str)
+        OPT_NOTIFY("%s: %s", en->c.en.name, map->str);
+      else
+        OPT_NOTIFY("%s: %d", en->c.en.name, map->val);
     }
   }
 }
@@ -354,13 +388,15 @@ static void arrow_cb(GtkToggleButton *self G_GNUC_UNUSED, t_ent_exp_common *en) 
   if (en && GTK_IS_TOGGLE_BUTTON(en->arrow)) {
     gboolean active = gtk_toggle_button_get_active(GTK_TOGGLE_BUTTON(en->arrow));
     g_object_set(G_OBJECT(en->arrow), ICON_PROP, active ? GO_UP_ICON : GO_DOWN_ICON, NULL);
-    if (GTK_IS_WIDGET(en->sub)) gtk_widget_set_visible(en->sub, active);
+    if (GTK_IS_WIDGET(en->sub))
+      gtk_widget_set_visible(en->sub, active);
     DEBUG("%s: %d", en->en.name, active);
   }
 }
 
 static void input_cb(GtkWidget *input, t_ent_str *en) {
-  if (!en) return;
+  if (!en)
+    return;
   g_return_if_fail(GTK_IS_EDITABLE(input));
   const char *got = gtk_editable_get_text(GTK_EDITABLE(input));
   switch (en->en.type) {
@@ -373,16 +409,20 @@ static void input_cb(GtkWidget *input, t_ent_str *en) {
       if (okay) {
         if (en->pint) {
           int n = *en->pint;
-          /*extra*/ if (en->en.type == ENT_STR_IVAL) opts.tout_usec = n * G_USEC_PER_SEC;
+          if (en->en.type == ENT_STR_IVAL) // extra
+            opts.tout_usec = n * G_USEC_PER_SEC;
           OPT_NOTIFY("%s: %d", en->en.name, n);
         }
-      } else set_ed_texthint(en);
-      if (en->en.type == ENT_STR_IVAL) drawtab_refresh(); // rescale time-axis
+      } else
+        set_ed_texthint(en);
+      if (en->en.type == ENT_STR_IVAL)
+        drawtab_refresh(); // rescale time-axis
     } break;
     case ENT_STR_PLOAD: {
       char *pad = parser_str(got, en->en.name, OPT_TYPE_PAD);
       if (pad) {
-        if (en->pstr) g_strlcpy(en->pstr, pad, en->slen);
+        if (en->pstr)
+          g_strlcpy(en->pstr, pad, en->slen);
         OPT_NOTIFY("%s: %s", en->en.name, pad);
         g_free(pad);
       } else set_ed_texthint(en);
@@ -393,11 +433,15 @@ static void input_cb(GtkWidget *input, t_ent_str *en) {
 }
 
 static void minttl_cb(GtkSpinButton *spin, t_ent_spn_elem *en) {
-  if (!GTK_IS_SPIN_BUTTON(spin) || !en) return;
-  int *pmin = en->aux[0].pval; if (!pmin) return;
+  if (!GTK_IS_SPIN_BUTTON(spin) || !en)
+    return;
+  int *pmin = en->aux[0].pval;
+  if (!pmin)
+    return;
   int got = gtk_spin_button_get_value_as_int(spin);
   int val = got - 1;
-  if (*pmin == val) return;
+  if (*pmin == val)
+    return;
   t_ent_spn_aux *aux = &en->aux[1];
   int *plim = aux->pval;
   const t_minmax *minmax = aux->mm ? aux->mm : &opt_mm_ttl;
@@ -417,10 +461,14 @@ static void minttl_cb(GtkSpinButton *spin, t_ent_spn_elem *en) {
 }
 
 static void maxttl_cb(GtkSpinButton *spin, t_ent_spn_elem *en) {
-  if (!GTK_IS_SPIN_BUTTON(spin) || !en) return;
-  int *plim = en->aux[1].pval; if (!plim) return;
+  if (!GTK_IS_SPIN_BUTTON(spin) || !en)
+    return;
+  int *plim = en->aux[1].pval;
+  if (!plim)
+    return;
   int got = gtk_spin_button_get_value_as_int(GTK_SPIN_BUTTON(spin));
-  if (*plim == got) return;
+  if (*plim == got)
+    return;
   t_ent_spn_aux *aux = &en->aux[0];
   int *pmin = aux->pval;
   int min = (pmin ? *pmin : 0) + 1;
@@ -441,10 +489,14 @@ static void maxttl_cb(GtkSpinButton *spin, t_ent_spn_elem *en) {
 
 #ifdef WITH_PLOT
 static void colgrad_cb(GtkSpinButton *spin, t_ent_spn_aux *aux) {
-  if (!GTK_IS_SPIN_BUTTON(spin) || !aux) return;
-  int *pval = aux->pval; if (!pval) return;
+  if (!GTK_IS_SPIN_BUTTON(spin) || !aux)
+    return;
+  int *pval = aux->pval;
+  if (!pval)
+    return;
   int got = gtk_spin_button_get_value_as_int(spin);
-  if (*pval == got) return;
+  if (*pval == got)
+    return;
   const t_minmax *minmax = aux->mm ? aux->mm : &opt_mm_col;
   if ((minmax->min <= got) && (got <= minmax->max)) {
     *pval = got;
@@ -454,22 +506,30 @@ static void colgrad_cb(GtkSpinButton *spin, t_ent_spn_aux *aux) {
     else
       OPT_NOTIFY("%s: %d (0x%02x)", aux->prfx, got, got);
     plottab_refresh(PL_PARAM_COLOR);
-  } else WARN("%s [%d,%d]: %d", OUTRANGE_ERR, minmax->min, minmax->max, got);
+  } else
+    WARN("%s [%d,%d]: %d", OUTRANGE_ERR, minmax->min, minmax->max, got);
 }
 
 void set_rotor_n_redraw(int step, gboolean rev, int n) {
-  if (!step) return;
+  if (!step)
+    return;
   int q[4] = { 0, 0, 0, rev ? -step : step }; q[n] = 1;
-  for (int i = 0; i < 3; i++) q[i] = (i == n);
-  plottab_on_trans_opts(q); plottab_redraw();
+  for (int i = 0; i < 3; i++)
+    q[i] = (i == n);
+  plottab_on_trans_opts(q);
+  plottab_redraw();
 }
 
 static void rotation_cb(GtkSpinButton *spin, t_ent_spn_aux *aux) {
-  if (!GTK_IS_SPIN_BUTTON(spin) || !aux) return;
+  if (!GTK_IS_SPIN_BUTTON(spin) || !aux)
+    return;
   int *pval = &aux->pval[aux->pn];
   int want = gtk_spin_button_get_value_as_int(spin);
-  int step = want - *pval; if (aux->wrap) step %= aux->wrap;
-  if (!step) return;
+  int step = want - *pval;
+  if (aux->wrap)
+    step %= aux->wrap;
+  if (!step)
+    return;
   const t_minmax *minmax = aux->mm ? aux->mm : &opt_mm_rot;
   if ((minmax->min <= want) && (want <= minmax->max)) {
     OPT_NOTIFY("%s: %d", aux->prfx, want);
@@ -477,17 +537,26 @@ static void rotation_cb(GtkSpinButton *spin, t_ent_spn_aux *aux) {
       *pval = want; t_ent_spn_elem *list = ent_spn[aux->sn].list;
       if (list) for (int i = 0; i < ENT_SPN_ANGS; i++) { // adjust angle entries' step
         GtkSpinButton *spin = list->aux[i].spin;
-        if (GTK_IS_SPIN_BUTTON(spin)) gtk_spin_button_set_increments(spin, want, want * 6); }
-    } else { *pval = want; set_rotor_n_redraw(step, aux->rev, aux->pn); }
+        if (GTK_IS_SPIN_BUTTON(spin))
+          gtk_spin_button_set_increments(spin, want, want * 6);
+      }
+    } else {
+      *pval = want;
+      set_rotor_n_redraw(step, aux->rev, aux->pn);
+    }
   }
 }
 
 static void scalefov_cb(GtkSpinButton *spin, t_ent_spn_aux *aux) {
-  if (!GTK_IS_SPIN_BUTTON(spin) || !aux) return;
-  int *pval = aux->pval; if (!pval) return;
+  if (!GTK_IS_SPIN_BUTTON(spin) || !aux)
+    return;
+  int *pval = aux->pval;
+  if (!pval)
+    return;
   int got = gtk_spin_button_get_value_as_int(spin);
   const t_minmax *minmax = aux->mm ? aux->mm : &opt_mm_fov;
-  if ((*pval == got) || (got < minmax->min) || (got > minmax->max)) return;
+  if ((*pval == got) || (got < minmax->min) || (got > minmax->max))
+    return;
   *pval = got;
   OPT_NOTIFY("%s: %d°", aux->prfx, got);
   plottab_refresh(PL_PARAM_FOV);
@@ -495,62 +564,68 @@ static void scalefov_cb(GtkSpinButton *spin, t_ent_spn_aux *aux) {
 #endif
 
 static void graph_type_cb(void) {
-  if (!pinger_state.pause) graphtab_update();
+  if (!pinger_state.pause)
+    graphtab_update();
   graphtab_refresh();
 }
 
 static GtkWidget* label_box(const char *name, const char *unit) {
   GtkWidget *box = gtk_box_new(GTK_ORIENTATION_HORIZONTAL, MARGIN * 2);
   g_return_val_if_fail(GTK_IS_BOX(box), NULL);
-  char *name_u = unit ?
-    g_strdup_printf("%s, %s", name, unit):
-    g_strdup(name);
-  GtkWidget *label = gtk_label_new(name_u);
-  g_free(name_u);
+  char *nameunit = unit ? g_strdup_printf("%s, %s", name, unit) : NULL;
+  GtkWidget *label = gtk_label_new(nameunit ? nameunit : name);
+  g_free(nameunit);
   g_return_val_if_fail(GTK_IS_LABEL(label), box);
   gtk_label_set_use_underline(GTK_LABEL(label), true);
   gtk_box_append(GTK_BOX(box), label);
-  if (style_loaded) gtk_widget_add_css_class(label, CSS_PAD);
+  if (style_loaded)
+    gtk_widget_add_css_class(label, CSS_PAD);
   gtk_widget_set_halign(label, GTK_ALIGN_START);
   gtk_widget_set_hexpand(label, true);
   return box;
 }
 
 static GtkWidget* add_labelbox(GtkWidget* list, const char *name, const char *unit) {
+  // list is either GtkBox or GtkListBox
   GtkWidget *box = label_box(name, unit);
   if (box) {
-    if (GTK_IS_LIST_BOX(list)) gtk_list_box_append(GTK_LIST_BOX(list), box);
-    else if (GTK_IS_BOX(list)) gtk_box_append(GTK_BOX(list), box);
+    if (GTK_IS_LIST_BOX(list))
+      gtk_list_box_append(GTK_LIST_BOX(list), box);
+    else if (GTK_IS_BOX(list))
+      gtk_box_append(GTK_BOX(list), box);
   }
   return box;
 }
 
-static GtkWidget* add_opt_check(GtkWidget* list, t_ent_bool *en) {
-  if (!en) return NULL;
+static GtkWidget* add_opt_check(GtkWidget* list, t_ent_bool *en) { // NONNULL(1, 2)
+  // list is either GtkBox or GtkListBox
   GtkWidget *box = add_labelbox(list, en->en.name, en->en.unit);
   if (box) {
     GtkWidget *check = gtk_check_button_new();
     g_return_val_if_fail(GTK_IS_CHECK_BUTTON(check), box);
     gtk_box_append(GTK_BOX(box), check);
-    if (style_loaded) gtk_widget_add_css_class(check, CSS_RPAD);
+    if (style_loaded)
+      gtk_widget_add_css_class(check, CSS_RPAD);
     gtk_widget_set_halign(check, GTK_ALIGN_END);
     g_signal_connect(check, EV_TOGGLE, G_CALLBACK(toggle_cb), en);
     en->check = GTK_CHECK_BUTTON(check);
     gboolean *pbool = EN_BOOLPTR(en);
-    if (pbool) gtk_check_button_set_active(en->check, *pbool);
+    if (pbool)
+      gtk_check_button_set_active(en->check, *pbool);
   }
   return box;
 }
 
-static GtkWidget* add_opt_enter(GtkWidget* list, t_ent_str *en) {
-  if (!en) return NULL;
+static GtkWidget* add_opt_enter(GtkWidget* list, t_ent_str *en) { // NONNULL(1, 2)
+  // list is GtkListBox
   GtkWidget *box = add_labelbox(list, en->en.name, en->en.unit);
   if (box) {
     en->box = box;
     en->input = gtk_entry_new();
     g_return_val_if_fail(GTK_IS_ENTRY(en->input), box);
     gtk_box_append(GTK_BOX(box), en->input);
-    if (style_loaded) gtk_widget_add_css_class(en->input, CSS_NOPAD);
+    if (style_loaded)
+      gtk_widget_add_css_class(en->input, CSS_NOPAD);
     gtk_widget_set_halign(en->input, GTK_ALIGN_END);
     gtk_entry_set_alignment(GTK_ENTRY(en->input), 0.99);
     gtk_entry_set_max_length(GTK_ENTRY(en->input), en->len);
@@ -562,8 +637,8 @@ static GtkWidget* add_opt_enter(GtkWidget* list, t_ent_str *en) {
   return box;
 }
 
-static GtkWidget* add_expand_common(GtkWidget* list, t_ent_exp_common *en) {
-  if (!en) return NULL;
+static GtkWidget* add_expand_common(GtkWidget* list, t_ent_exp_common *en) { // NONNULL(1, 2)
+  // list is GtkListBox
   GtkWidget *box = gtk_box_new(GTK_ORIENTATION_VERTICAL, 0);
   g_return_val_if_fail(GTK_IS_BOX(box), NULL);
   en->box = box;
@@ -575,7 +650,8 @@ static GtkWidget* add_expand_common(GtkWidget* list, t_ent_exp_common *en) {
     en->arrow = gtk_toggle_button_new();
     g_return_val_if_fail(GTK_IS_TOGGLE_BUTTON(en->arrow), box);
     gtk_box_append(GTK_BOX(title), en->arrow);
-    if (style_loaded) gtk_widget_add_css_class(en->arrow, CSS_NOFRAME);
+    if (style_loaded)
+      gtk_widget_add_css_class(en->arrow, CSS_NOFRAME);
     gtk_widget_set_halign(en->arrow, GTK_ALIGN_END);
     g_object_set(G_OBJECT(en->arrow), ICON_PROP, GO_DOWN_ICON, NULL);
     g_signal_connect(en->arrow, EV_TOGGLE, G_CALLBACK(arrow_cb), en);
@@ -588,16 +664,31 @@ static GtkWidget* add_expand_common(GtkWidget* list, t_ent_exp_common *en) {
   return box;
 }
 
-static GtkWidget* add_opt_expand(GtkWidget* list, t_ent_exp *en) {
-  if (!en) return NULL;
-  GtkWidget *box = add_expand_common(list, &en->c);
-  if (GTK_IS_BOX(box) && en->c.sub && en->desc) {
-    t_elem_desc *desc = en->desc; type2ent_fn type2ent = desc->t2e;
-    if (type2ent) for (int i = desc->mm.min; i <= desc->mm.max; i++) {
-      int type = desc->elems[i].type;
-      int ndx = type2ent(type);
-      if (ndx < 0) WARN("%s: %d", UNKNTYPE_HDR, type);
-      else add_opt_check(en->c.sub, &ent_bool[ndx]);
+static void add_opt_exp_entry(GtkWidget *list, t_elem_desc *desc) { // NONNULL(1, 2)
+  // list is GtkListBox
+  type2ent_fn type2ent = desc->t2e;
+  if (type2ent) for (int i = desc->mm.min; i <= desc->mm.max; i++) {
+    int type = desc->elems[i].type;
+    int ndx = type2ent(type);
+    if (ndx < 0)
+      WARN("%s: %d", UNKNTYPE_HDR, type);
+    else
+      add_opt_check(list, &ent_bool[ndx]);
+  }
+}
+
+static GtkWidget* add_opt_expand(GtkWidget* list, t_ent_exp ens[], uint en_len) { // NONNULL(1, 2)
+  // list is GtkListBox
+  if (en_len < 1)
+    return NULL;
+  GtkWidget *box = add_expand_common(list, &ens->c);
+  if (GTK_IS_BOX(box)) {
+    GtkWidget *sub = ens->c.sub;
+    if (sub) {
+      t_ent_exp *en = ens;
+      for (uint i = 0; i < en_len; i++, en++)
+        if (en->desc)
+          add_opt_exp_entry(sub, en->desc);
     }
   }
   return box;
@@ -606,7 +697,7 @@ static GtkWidget* add_opt_expand(GtkWidget* list, t_ent_exp *en) {
 #define SPN_MM_OR_DEF(aux, inc) ((aux).pval ? (*((aux).pval) + (inc)) : (aux).def)
 #define SPN_MM_NDX(mm_ndx, inc) ((ndx == (mm_ndx)) ? (en->aux[mm_ndx].def) : SPN_MM_OR_DEF(en->aux[mm_ndx], inc))
 
-static gboolean add_minmax(GtkWidget *box, t_ent_spn_elem *en, guint ndx, const int *step, gboolean wrap) {
+static gboolean add_minmax(GtkWidget *box, t_ent_spn_elem *en, uint ndx, const int *step, gboolean wrap) {
   gboolean okay = false;
   if (GTK_IS_BOX(box) && en) {
     const t_minmax *mm0 = en->aux[en->bond ? 0 : ndx].mm;
@@ -621,7 +712,8 @@ static gboolean add_minmax(GtkWidget *box, t_ent_spn_elem *en, guint ndx, const 
         int *pval = &en->aux[ndx].pval[pn];
         if (pval) {
           int val = *pval;
-          if (en->bond && !ndx) val++;
+          if (en->bond && !ndx)
+            val++;
           gtk_spin_button_set_value(GTK_SPIN_BUTTON(spin), val);
         }
         gtk_spin_button_set_wrap(GTK_SPIN_BUTTON(spin), wrap);
@@ -630,24 +722,27 @@ static gboolean add_minmax(GtkWidget *box, t_ent_spn_elem *en, guint ndx, const 
         if (en->aux[ndx].cb)
           g_signal_connect(spin, EV_VAL_CHANGE, en->aux[ndx].cb, en->bond ? (void*)en : &en->aux[ndx]);
         okay = true;
-     } else g_warning("%s: %s", ERROR_HDR, "gtk_spin_button_new_with_range()");
-  }}
+      } else
+        g_warning("%s: %s", ERROR_HDR, "gtk_spin_button_new_with_range()");
+    }
+  }
   return okay;
 }
 
 static void grey_into_box(GtkWidget *box, GtkWidget *widget, gboolean sens, gboolean hexp) {
-  if (!GTK_IS_BOX(box) || !GTK_IS_WIDGET(widget)) return;
+  if (!GTK_IS_BOX(box) || !GTK_IS_WIDGET(widget))
+    return;
   gtk_widget_set_sensitive(widget, sens);
   gtk_widget_set_hexpand(widget, hexp);
   gtk_box_append(GTK_BOX(box), widget);
 }
 
-static GtkWidget* add_opt_range(GtkWidget* listbox, t_ent_spn *en) {
-  if (!GTK_LIST_BOX(listbox) || !en) return NULL;
-  GtkWidget *box = add_expand_common(listbox, &en->c);
+static GtkWidget* add_opt_range(GtkWidget* list, t_ent_spn *en) { // NONNULL(1, 2)
+  // list is GtkListBox
+  GtkWidget *box = add_expand_common(list, &en->c);
   if (GTK_IS_BOX(box) && en->c.sub) {
     t_ent_spn_elem *elem = en->list;
-    for (guint n = 0; (n < G_N_ELEMENTS(en->list)) && elem && elem->title; n++, elem++) {
+    for (uint n = 0; (n < G_N_ELEMENTS(en->list)) && elem && elem->title; n++, elem++) {
       GtkWidget *subbox = NULL;
       switch (elem->kind) {
 #ifdef WITH_PLOT
@@ -655,7 +750,8 @@ static GtkWidget* add_opt_range(GtkWidget* listbox, t_ent_spn *en) {
 #endif
         default: subbox = gtk_box_new(GTK_ORIENTATION_HORIZONTAL, MARGIN); break;
       }
-      if (!GTK_IS_BOX(subbox)) continue;
+      if (!GTK_IS_BOX(subbox))
+        continue;
       gtk_list_box_append(GTK_LIST_BOX(en->c.sub), subbox);
       switch (elem->kind) {
         case MINIMAX_SPIN:
@@ -668,16 +764,20 @@ static GtkWidget* add_opt_range(GtkWidget* listbox, t_ent_spn *en) {
 #ifdef WITH_PLOT
         case ROTOR_COLUMN:
           add_opt_check(subbox, &ent_bool[ENT_BOOL_GLOB]);
-          for (guint i = 0; i < ENT_SPN_ROT_MAX; i++) if (elem->aux[i].prfx) {
+          for (uint i = 0; i < ENT_SPN_ROT_MAX; i++) if (elem->aux[i].prfx) {
             GtkWidget *row = gtk_box_new(GTK_ORIENTATION_HORIZONTAL, MARGIN);
-            if (!row) continue;
+            if (!row)
+              continue;
             grey_into_box(row, gtk_label_new(elem->aux[i].prfx), true, true);
             add_minmax(row, elem, i, elem->aux[i].step, elem->aux[i].wrap);
             gtk_box_append(GTK_BOX(subbox), row);
-            if (/*(i >= ENT_SPN_ROT0) &&*/ i <= ENT_SPN_ROT2) // leave only buttons for rotation spins
+            if (/*(i >= ENT_SPN_ROT0) &&*/ i <= ENT_SPN_ROT2)
+              // leave only buttons for rotation spins
               for (GtkWidget *txt = gtk_widget_get_first_child(GTK_WIDGET(elem->aux[i].spin));
-                GTK_IS_WIDGET(txt); txt = gtk_widget_get_next_sibling(txt))
-                  if (GTK_IS_TEXT(txt)) gtk_widget_set_visible(txt, false);
+                   GTK_IS_WIDGET(txt);
+                   txt = gtk_widget_get_next_sibling(txt))
+                if (GTK_IS_TEXT(txt))
+                  gtk_widget_set_visible(txt, false);
           }
           break;
         case SCALE_SPIN:
@@ -692,19 +792,25 @@ static GtkWidget* add_opt_range(GtkWidget* listbox, t_ent_spn *en) {
   return box;
 }
 
-static GtkWidget* add_opt_radio(GtkWidget* list, t_ent_rad *en) {
-  if (!en) return NULL;
+static GtkWidget* add_opt_radio(GtkWidget* list, t_ent_rad *en) { // NONNULL(1, 2)
+  // list is GtkListBox
+  if (!en)
+    return NULL;
   GtkWidget *box = add_expand_common(list, &en->c);
   if (GTK_IS_BOX(box) && en->c.sub) {
     GtkWidget *group = NULL;
     t_ent_rad_map *map = en->map;
-    for (guint i = 0; (i < G_N_ELEMENTS(en->map)) && map && map->str; i++, map++) {
+    for (uint i = 0; (i < G_N_ELEMENTS(en->map)) && map && map->str; i++, map++) {
       GtkWidget *button = gtk_check_button_new_with_label(map->str);
       g_return_val_if_fail(GTK_CHECK_BUTTON(button), box);
-      if (style_loaded) gtk_widget_add_css_class(button, CSS_PAD);
-      if (group) gtk_check_button_set_group(GTK_CHECK_BUTTON(button), GTK_CHECK_BUTTON(group));
-      else group = button;
-      if (en->pval) gtk_check_button_set_active(GTK_CHECK_BUTTON(button), *en->pval == map->val);
+      if (style_loaded)
+        gtk_widget_add_css_class(button, CSS_PAD);
+      if (group)
+        gtk_check_button_set_group(GTK_CHECK_BUTTON(button), GTK_CHECK_BUTTON(group));
+      else
+        group = button;
+      if (en->pval)
+        gtk_check_button_set_active(GTK_CHECK_BUTTON(button), *en->pval == map->val);
       g_signal_connect(button, EV_TOGGLE, G_CALLBACK(radio_cb), map);
       gtk_list_box_append(GTK_LIST_BOX(en->c.sub), button);
     }
@@ -717,7 +823,8 @@ static void gtk_compat_list_box_remove_all(GtkListBox *box) {
 #if GTK_CHECK_VERSION(4, 12, 0)
   gtk_list_box_remove_all(box);
 #else
-  for (GtkWidget *c; (c = gtk_widget_get_first_child(GTK_WIDGET(box)));) gtk_list_box_remove(box, c);
+  for (GtkWidget *c; (c = gtk_widget_get_first_child(GTK_WIDGET(box)));)
+    gtk_list_box_remove(box, c);
 #endif
 }
 
@@ -728,9 +835,12 @@ static gboolean create_optmenu(GtkWidget *bar, const char **icons, const char *t
   gboolean okay = true;
   gtk_header_bar_pack_start(GTK_HEADER_BAR(bar), menu);
   const char *ico = is_sysicon(icons);
-  if (!ico) g_warning("%s: %s", OPTIONS_HDR, NOICON_ERR);
-  else gtk_menu_button_set_icon_name(GTK_MENU_BUTTON(menu), ico);
-  if (tooltip) gtk_widget_set_tooltip_text(menu, tooltip);
+  if (ico)
+    gtk_menu_button_set_icon_name(GTK_MENU_BUTTON(menu), ico);
+  else
+    g_warning("%s: %s", OPTIONS_HDR, NOICON_ERR);
+  if (tooltip)
+    gtk_widget_set_tooltip_text(menu, tooltip);
   GtkWidget *popover = gtk_popover_new();
   if (GTK_IS_POPOVER(popover)) {
     gtk_menu_button_set_popover(GTK_MENU_BUTTON(menu), popover);
@@ -741,40 +851,55 @@ static gboolean create_optmenu(GtkWidget *bar, const char **icons, const char *t
       gtk_list_box_set_selection_mode(GTK_LIST_BOX(list), GTK_SELECTION_NONE);
       gtk_widget_set_halign(list, GTK_ALIGN_FILL);
       gtk_widget_set_hexpand(list, false);
-      if (fn) okay = fn(list);
-    } else okay = false;
-  } else okay = false;
-  if (!okay) gtk_header_bar_remove(GTK_HEADER_BAR(bar), menu);
+      if (fn)
+        okay = fn(list);
+    } else
+      okay = false;
+  } else
+    okay = false;
+  if (!okay)
+    gtk_header_bar_remove(GTK_HEADER_BAR(bar), menu);
   return okay;
 }
 
-#define EN_PR_FMT(ndx, fmt, arg) { snprintg(ent_str[ndx].hint, sizeof(ent_str[ndx].hint), fmt, arg); \
-  if (!add_opt_enter(list, &ent_str[ndx])) okay = false; }
+#define EN_PR_FMT(ndx, fmt, arg) do {                               \
+  snprintg(ent_str[ndx].hint, sizeof(ent_str[ndx].hint), fmt, arg); \
+  if (!add_opt_enter(list, &ent_str[ndx]))                          \
+    okay = false;                                                   \
+} while (0)
 #define EN_PR_STR(ndx) EN_PR_FMT(ndx, "%s", ent_str[ndx].sdef)
 #define EN_PR_INT(ndx) EN_PR_FMT(ndx, "%d", ent_str[ndx].idef)
 
 static gboolean create_main_optmenu(GtkWidget *list) {
   static GtkWidget *main_optmenu_list;
-  if (list) main_optmenu_list = list; else list = main_optmenu_list;
+  if (list)
+    main_optmenu_list = list;
+  else
+    list = main_optmenu_list;
   g_return_val_if_fail(GTK_IS_LIST_BOX(list), false);
   gboolean okay = true;
   gtk_compat_list_box_remove_all(GTK_LIST_BOX(list));
   EN_PR_INT(ENT_STR_CYCLES);
   EN_PR_INT(ENT_STR_IVAL);
-  if (!add_opt_check(list,  &ent_bool[ENT_BOOL_DNS])) okay = false;
-  if (!add_opt_expand(list, &ent_exp[ENT_EXP_INFO]))  okay = false;
-  if (!add_opt_expand(list, &ent_exp[ENT_EXP_STAT]))  okay = false;
-  if (!add_opt_range(list,  &ent_spn[ENT_SPN_TTL]))   okay = false;
+  if (!add_opt_check(list,  &ent_bool[ENT_BOOL_DNS]))   okay = false;
+  // adjacent EXP_INFO and EXP_WMF: len = 2
+  if (!add_opt_expand(list, &ent_exp[ENT_EXP_INFO], 2)) okay = false;
+  //
+  if (!add_opt_expand(list, &ent_exp[ENT_EXP_STAT], 1)) okay = false;
+  if (!add_opt_range(list,  &ent_spn[ENT_SPN_TTL]))     okay = false;
   EN_PR_INT(ENT_STR_QOS);
   EN_PR_STR(ENT_STR_PLOAD);
   EN_PR_INT(ENT_STR_PSIZE);
-  if (!add_opt_radio(list, &ent_rad[ENT_RAD_IPV]))    okay = false;
+  if (!add_opt_radio(list, &ent_rad[ENT_RAD_IPV]))      okay = false;
   return okay;
 }
 
 static gboolean create_ext_optmenu(GtkWidget *list) {
   static GtkWidget *ext_optmenu_list;
-  if (list) ext_optmenu_list = list; else list = ext_optmenu_list;
+  if (list)
+    ext_optmenu_list = list;
+  else
+    list = ext_optmenu_list;
   g_return_val_if_fail(GTK_IS_LIST_BOX(list), false);
   gboolean okay = true;
   gtk_compat_list_box_remove_all(GTK_LIST_BOX(list));
@@ -787,11 +912,11 @@ static gboolean create_ext_optmenu(GtkWidget *list) {
 #endif
     if (!add_opt_radio(list,    &ent_rad[ENT_RAD_GRAPH]))     okay = false;
     if (!add_opt_check(list,    &ent_bool[ENT_BOOL_LGND]))    okay = false;
-    if (!add_opt_expand(list,   &ent_exp[ENT_EXP_LGFL]))      okay = false;
-    if (!add_opt_expand(list,   &ent_exp[ENT_EXP_GREX]))      okay = false;
+    if (!add_opt_expand(list,   &ent_exp[ENT_EXP_LGFL], 1))   okay = false;
+    if (!add_opt_expand(list,   &ent_exp[ENT_EXP_GREX], 1))   okay = false;
 #ifdef WITH_PLOT
     if (opts.plot) {
-      if (!add_opt_expand(list, &ent_exp[ENT_EXP_PLEL]))      okay = false;
+      if (!add_opt_expand(list, &ent_exp[ENT_EXP_PLEL], 1))   okay = false;
       if (!add_opt_range(list,  &ent_spn[ENT_SPN_COLOR]))     okay = false;
       int rotent = opts.rglob ? ENT_SPN_GLOBAL : ENT_SPN_LOCAL;
       if (!add_opt_range(list,  &ent_spn[rotent]))            okay = false;
@@ -845,6 +970,7 @@ void init_option_links(void) {
   INIT_BE_INFO(ENT_BOOL_CC,   ELEM_CC_HEADER  );
   INIT_BE_INFO(ENT_BOOL_DESC, ELEM_DESC_HEADER);
   INIT_BE_INFO(ENT_BOOL_RT,   ELEM_RT_HDR     );
+  INIT_BE_INFO(ENT_BOOL_WMF,  ELEM_WMF_HDR    );
   INIT_BE_STAT(ENT_BOOL_LOSS, ELEM_LOSS_HEADER);
   INIT_BE_STAT(ENT_BOOL_SENT, ELEM_SENT_HEADER);
   INIT_BE_STAT(ENT_BOOL_RECV, ELEM_RECV_HEADER);
@@ -873,7 +999,8 @@ void init_option_links(void) {
   ent_bool[ENT_BOOL_GLOB].en.name = OPT_GLOB_HDR;
 #endif
   //
-  INIT_EE_ND(ENT_EXP_INFO, OPT_INFO_HDR, info_desc);
+  INIT_EE_ND(ENT_EXP_INFO, OPT_INFO_HDR, info_desc); // whois
+  INIT_EE_ND(ENT_EXP_WMF,  OPT_INFO_HDR, wmf_desc);  // whois multifield
   INIT_EE_ND(ENT_EXP_STAT, OPT_STAT_HDR, stat_desc);
   INIT_EE_ND(ENT_EXP_LGFL, OPT_GRLG_HDR, grlg_desc);
   INIT_EE_ND(ENT_EXP_GREX, OPT_GREX_HDR, grex_desc);
@@ -936,20 +1063,21 @@ gboolean option_init(GtkWidget* bar) {
   return true;
 }
 
-#define ENT_LOOP_SET(row) do {                                   \
-  if (GTK_IS_WIDGET(row)) gtk_widget_set_sensitive(row, notrun); \
+#define ENT_LOOP_SET(row) do {             \
+  if (GTK_IS_WIDGET(row))                  \
+    gtk_widget_set_sensitive(row, notrun); \
 } while (0)
-#define ENT_LOOP_STR(arr) do {                  \
-  for (guint i = 0; i < G_N_ELEMENTS(arr); i++) \
-    ENT_LOOP_SET((arr)[i].box);                 \
+#define ENT_LOOP_STR(arr) do {                 \
+  for (uint i = 0; i < G_N_ELEMENTS(arr); i++) \
+    ENT_LOOP_SET((arr)[i].box);                \
 } while (0)
 #define ENT_LOOP_EXP(arr) do {                    \
-  for (guint i = 0; i < G_N_ELEMENTS(arr); i++)   \
+  for (uint i = 0; i < G_N_ELEMENTS(arr); i++)    \
     if (!(arr)[i].c.atrun) {                      \
       ENT_LOOP_SET((arr)[i].c.box);               \
       if (GTK_IS_TOGGLE_BUTTON((arr)[i].c.arrow)) \
         gtk_toggle_button_set_active(GTK_TOGGLE_BUTTON((arr)[i].c.arrow), false); \
-    } \
+    }                                             \
 } while (0)
 
 void option_update(void) {
@@ -959,13 +1087,27 @@ void option_update(void) {
   ENT_LOOP_EXP(ent_rad);
 }
 
-void option_up_menu_main(void) { if (!opt_silent) { opt_silent = true; create_main_optmenu(NULL); opt_silent = false; }}
-void option_up_menu_ext(void)  { if (!opt_silent) { opt_silent = true; create_ext_optmenu(NULL);  opt_silent = false; }}
+void option_up_menu_main(void) {
+  if (!opt_silent) {
+    opt_silent = true;
+    create_main_optmenu(NULL);
+    opt_silent = false;
+  }
+}
+
+void option_up_menu_ext(void) {
+  if (!opt_silent) {
+    opt_silent = true;
+    create_ext_optmenu(NULL);
+    opt_silent = false;
+  }
+}
 
 void option_legend(void) {
   t_ent_bool *en = &ent_bool[ENT_BOOL_LGND];
   gboolean *pbool = EN_BOOLPTR(en);
-  if (en->check && pbool) gtk_check_button_set_active(en->check, *pbool);
-  toggle_legend();
+  if (en->check && pbool)
+    gtk_check_button_set_active(en->check, *pbool);
+  toggled_legend();
 }